requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
requests>=2.31
python-dotenv==1.0.1
supabase==2.9.1
httpx==0.27.2
mangum==0.17.0
